переделано добавление розетки. Вынесены методы для переиспользования
