+++ src/test/java/javaTests/steps/AddDeviceSteps.java
@@ -1,7 +1,5 @@
 package javaTests.steps;
 
-import org.testng.Assert;
-
 import io.appium.java_client.android.AndroidDriver;
 import io.qameta.allure.Step;
 import javaTests.pageObjects.AddDevicePage;
@@ -20,46 +18,10 @@ public class AddDeviceSteps {
         addDevicePage.waitForElementToClick(addDevicePage.addDeviceBtn).click();
     }
 
-    @Step("Нажимаем кнопку 'Далее'")
-    public void clickNextBtn() {
-        addDevicePage.waitForElementToClick(addDevicePage.nextBtn).click();
-    }
-
     @Step("Нажимаем добавить розетку")
     public void clickAddWifiSocket() {
         addDevicePage.waitForElementToClick(addDevicePage.powerSupplyGroupExpandableList).click();
         addDevicePage.waitForElementToClick(addDevicePage.wifiSocketExpandableList).click();
-    }
-
-    //Action with field:
-    @Step("Нажимаем кнопку 'Вижу'")
-    public void clickViewBtn() {
-        addDevicePage.waitForElementToClick(addDevicePage.viewBtn).click();
-    }
-
-    @Step("Нажимаем на справку")
-    public void clickHelpBtn() throws InterruptedException {
-        addDevicePage.waitForElementToClick(addDevicePage.helpBtn).click();
-        Thread.sleep(2000);
-    }
-
-    //Check text:
-    @Step("Текст кнопки 'вижу' присуствует на кнопке")
-    void checkViewBtnText() {
-        String textBtn = addDevicePage.waitForElementIsDisplayed(addDevicePage.viewBtn).getText();
-        Assert.assertEquals(textBtn, "Вижу");
-    }
-
-    @Step("Заголовок 'Подключение' присуствует на экране смартконфига устройств")
-    void checkPrepareDeviceSmartConfigTitle() {
-        String title = addDevicePage.waitForElementIsDisplayed(addDevicePage.addDeviceAllScreenTitle).getText();
-        Assert.assertEquals(title, "Подключение");
-    }
-
-    @Step("Текст присутствует на кнопке 'далее'")
-    void checkNextBtnText() {
-        String textBtn = addDevicePage.waitForElementIsDisplayed(addDevicePage.nextBtn).getText();
-        Assert.assertEquals(textBtn, "Далее");
     }
 
     //Other checks:
@@ -67,7 +29,7 @@ public class AddDeviceSteps {
     public void checkAddDeviceGroupEmptyHouse() throws InterruptedException {
 
         checkBackBtn();
-        addDevicePage.waitForElementToClick(addDevicePage.addDeviceAllScreenTitle).isDisplayed();
+        addDevicePage.waitForElementToClick(addDevicePage.addDeviceScreenTitle).isDisplayed();
         Thread.sleep(4000);
         addDevicePage.controlGroupExpandableList.click();
         addDevicePage.waitForElementToClick(addDevicePage.controlModuleExpandableList).isDisplayed();
@@ -103,35 +65,11 @@ public class AddDeviceSteps {
     }
 
     @Step("Кнопка назад присутствует")
-    void checkBackBtn() {
+    private void checkBackBtn() {
         addDevicePage.waitForElementToClick(addDevicePage.backBtn).isDisplayed();
         addDevicePage.backBtn.isEnabled();
     }
 
-    @Step("Кнопка 'Вижу' не доступна")
-    public void checkViewBtnIsNotEnable() {
-        boolean enable = addDevicePage.waitForElementIsDisplayed(addDevicePage.viewBtn).isEnabled();
-        Assert.assertFalse(enable);
-    }
-
-    @Step("Кнопка 'Вижу' доступна")
-    void checkViewBtnIsEnable() throws InterruptedException {
-        Thread.sleep(3000);
-        boolean enable = addDevicePage.waitForElementIsDisplayed(addDevicePage.viewBtn).isEnabled();
-        Assert.assertTrue(enable);
-    }
-
-    @Step("Лодер появился. Старт смартконфига")
-    public void smartConfigIsStarted() {
-        boolean visibility = addDevicePage.longWaitForElementIsDisplayed(addDevicePage.progressLoader).isEnabled();
-        Assert.assertTrue(visibility);
-    }
-
-    @Step("Лодер присутствует в течении смартконфига")
-    void smartConfigLoaderIsEnabled() {
-        boolean enable = addDevicePage.longWaitForElementIsDisplayed(addDevicePage.progressLoader).isEnabled();
-        Assert.assertTrue(enable);
-    }
 }
