+++ src/test/java/javaTests/steps/DeviceListSteps.java
@@ -1,17 +1,18 @@
 package javaTests.steps;
 
-import org.testng.Assert;
-
 import io.appium.java_client.android.AndroidDriver;
 import io.qameta.allure.Step;
 import javaTests.pageObjects.DeviceListPage;
+import org.testng.Assert;
 
 public class DeviceListSteps {
 
     private final DeviceListPage deviceListPage;
+    private final RoomSteps room;
 
     public DeviceListSteps(AndroidDriver androidDriver) {
         deviceListPage = new DeviceListPage(androidDriver);
+        room = new RoomSteps(androidDriver);
     }
 
     @Step("Текст 'устройств пока нет' присутствует в пустом доме")
@@ -20,4 +21,19 @@ public class DeviceListSteps {
         Assert.assertEquals(text, "Устройств пока нет");
     }
 
+    @Step("Разворачиваем комнату Wifi")
+    private void rollWifiRoom() {
+        deviceListPage.waitForElementIsDisplayed(deviceListPage.roomWifiRoom).click();
+    }
+
+    @Step("Розетка отображается в списке устройств")
+    public boolean checkWifiSocketInDeviceList() {
+        try {
+            return deviceListPage.waitForElementIsDisplayed(deviceListPage.wifiSocketName).isDisplayed();
+        } catch (Exception e) {
+            rollWifiRoom();
+            deviceListPage.waitForElementIsDisplayed(deviceListPage.wifiSocketName).isDisplayed();
+            return false;
+        }
+    }
 }
